# backend/services/product_service/docs/swagger.yaml
basePath: /
definitions:
...
  product_service_internal_core_port_service_dto.AttributeRes:
    properties:
      id:
        type: integer
      name:
        type: string
      values:
        items:
          $ref: '#/definitions/product_service_internal_core_port_service_dto.AttributeValueRes'
        type: array
    type: object
  product_service_internal_core_port_service_dto.AttributeValueRes:
    properties:
      id:
        type: integer
      value:
        type: string
    type: object
...
  product_service_internal_core_port_service_dto.CreateAttributeReq:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
  product_service_internal_core_port_service_dto.CreateAttributeValueReq:
    properties:
      attribute_id:
        type: integer
      value:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - attribute_id
    - value
    type: object
...
  product_service_internal_core_port_service_dto.UpdateAttributeReq:
    properties:
      attribute_id:
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - attribute_id
    - name
    type: object
...
host: localhost:3002
info:
  contact:
    email: your.email@example.com
    name: API Support
  description: This is the Product Service API using Hexagonal Architecture & CQRS.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Product Service API
  version: "1.0"
paths:
  /attributes:
    get:
      description: Returns all attributes (e.g. Color, Size) each populated with their
        possible values.
      produces:
      - application/json
      responses:
        "200":
          description: Attribute list
          schema:
            items:
              $ref: '#/definitions/product_service_internal_core_port_service_dto.AttributeRes'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List all attributes with their values
      tags:
      - Attributes (Query)
  /attributes/{id}:
    get:
      description: Returns a single attribute with all its values.
      parameters:
      - description: Attribute ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Attribute detail
          schema:
            $ref: '#/definitions/product_service_internal_core_port_service_dto.AttributeRes'
        "404":
          description: Attribute not found
          schema:
            additionalProperties: true
            type: object
      summary: Get attribute by ID
      tags:
      - Attributes (Query)
  /attributes/admin:
    post:
      consumes:
      - application/json
      description: Creates a new attribute type such as Color or Size.
      parameters:
      - description: Attribute data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/product_service_internal_core_port_service_dto.CreateAttributeReq'
      produces:
      - application/json
      responses:
        "201":
          description: Attribute created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new attribute (Admin)
      tags:
      - Attributes (Command)
  /attributes/admin/{id}:
    delete:
      description: Deletes an attribute and cascades to its values.
      parameters:
      - description: Attribute ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Attribute deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Attribute not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete an attribute (Admin)
      tags:
      - Attributes (Command)
    put:
      consumes:
      - application/json
      description: Updates the name of an existing attribute.
      parameters:
      - description: Attribute ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated attribute data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/product_service_internal_core_port_service_dto.UpdateAttributeReq'
      produces:
      - application/json
      responses:
        "200":
          description: Attribute updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Attribute not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an attribute (Admin)
      tags:
      - Attributes (Command)
  /attributes/admin/{id}/values:
    post:
      consumes:
      - application/json
      description: Adds a new possible value (e.g. "Red") to an existing attribute
        (e.g. "Color").
      parameters:
      - description: Attribute ID
        in: path
        name: id
        required: true
        type: integer
      - description: Attribute value data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/product_service_internal_core_port_service_dto.CreateAttributeValueReq'
      produces:
      - application/json
      responses:
        "201":
          description: Attribute value created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Attribute not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add a value to an attribute (Admin)
      tags:
      - Attributes (Command)
  /attributes/admin/{id}/values/{valueId}:
    delete:
      description: Removes a specific value from an attribute.
      parameters:
      - description: Attribute ID
        in: path
        name: id
        required: true
        type: integer
      - description: Attribute Value ID
        in: path
        name: valueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Attribute value deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete an attribute value (Admin)
      tags:
      - Attributes (Command)
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer " followed by your token. Example: "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"